Vendor locales
So that external consumers don’t get the exotic subdependency warning when using pnpm
