+++ package.json
@@ -56,7 +56,7 @@
   },
   "scripts": {
     "dev": "next dev",
-    "build": "pnpm extract-locales && next build",
+    "build": "next build",
     "build:analyze": "npx next experimental-analyze",
     "start": "next start",
     "lint": "eslint .",
@@ -64,10 +64,9 @@
     "preinstall": "npx only-allow pnpm",
     "copy-locales": "cpx \"public/locales/**/*\" dist/locales",
     "copy-fonts": "cpx \"public/fonts/**/*\" dist/fonts",
-    "extract-locales": "node scripts/extract-locales.mjs",
     "check-locales": "node scripts/check-locales.mjs",
-    "update-locales": "pnpm remove @edrlab/thorium-locales && pnpm add @edrlab/thorium-locales@github:edrlab/thorium-locales && pnpm extract-locales && pnpm check-locales --summary --show-missing",
-    "bundle": "pnpm extract-locales && tsup --config tsup.config.ts && pnpm copy-locales && pnpm copy-fonts",
+    "update-locales": "node scripts/update-locales.mjs && pnpm check-locales --summary --show-missing",
+    "bundle": "tsup --config tsup.config.ts && pnpm copy-locales && pnpm copy-fonts",
     "increment": "node scripts/increment.mjs"
   },
   "peerDependencies": {
@@ -91,7 +90,6 @@
     "react-stately": "^3.44.0"
   },
   "dependencies": {
-    "@edrlab/thorium-locales": "github:edrlab/thorium-locales",
     "@readium/css": "^2.0.5",
     "@readium/navigator": "^2.8.0",
     "@readium/navigator-html-injectables": "^2.6.3",
