+++ pnpm-workspace.yaml
@@ -7,3 +7,8 @@ minimumReleaseAgeExclude:
   - "@readium/*"
 blockExoticSubdeps: true
 trustPolicy: no-downgrade
+# undici's 5.x maintenance line was never wired into npm provenance attestation,
+# unlike its 6.x/7.x mainline releases from the same maintainer (matteo.collina) -
+# a publish-pipeline gap, not a trust downgrade.
+trustPolicyExclude:
+  - "undici"
